requirements.txt
albumentations==0.4.5
attrdict==2.0.1
fire==0.3.1
jupyterlab==2.0.0
numpy==1.18.2
opencv-python-headless==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
pytorch-lightning==0.7.3
requests==2.23.0
scikit-learn==0.22.2.post1
tensorboard==2.2.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
